refactor: change package structure and code refactoring
