--- src/main/java/nextstep/auth/oauth2/OAuth2UserService.java
@@ -1,7 +1,0 @@
-package nextstep.auth.oauth2;
-
-public interface OAuth2UserService {
-
-    OAuth2User loadUser(OAuth2UserRequest oAuth2UserRequest);
-
-}
