+++ src/main/java/nextstep/auth/oauth2/dto/OAuth2User.java
@@ -1,7 +1,9 @@
-package nextstep.auth.oauth2;
+package nextstep.auth.oauth2.dto;
 
 public interface OAuth2User {
+
     String getUsername();
 
     String getRole();
+
 }
