+++ src/main/java/nextstep/auth/oauth2/dto/OAuth2UserRequest.java
@@ -1,4 +1,4 @@
-package nextstep.auth.oauth2;
+package nextstep.auth.oauth2.dto;
 
 public interface OAuth2UserRequest {
     String getUsername();
