+++ src/main/java/nextstep/auth/oauth2/github/dto/GithubProfileResponse.java
@@ -3,7 +3,7 @@ package nextstep.auth.oauth2.github.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import nextstep.auth.oauth2.OAuth2UserRequest;
+import nextstep.auth.oauth2.dto.OAuth2UserRequest;
 
 @NoArgsConstructor
 @AllArgsConstructor
