+++ src/main/java/nextstep/auth/token/service/TokenService.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import nextstep.global.error.code.ErrorCode;
 import nextstep.global.error.exception.AuthenticationException;
 import nextstep.auth.token.dto.TokenResponse;
-import nextstep.auth.oauth2.OAuth2User;
-import nextstep.auth.oauth2.OAuth2UserService;
+import nextstep.auth.oauth2.dto.OAuth2User;
+import nextstep.auth.oauth2.service.OAuth2UserService;
 import nextstep.auth.oauth2.github.client.GithubClient;
 import nextstep.auth.oauth2.github.dto.GithubProfileResponse;
 import nextstep.auth.userdetails.UserDetails;
