+++ src/main/java/nextstep/member/service/CustomOAuth2UserService.java
@@ -1,9 +1,9 @@
 package nextstep.member.service;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.auth.oauth2.OAuth2User;
-import nextstep.auth.oauth2.OAuth2UserRequest;
-import nextstep.auth.oauth2.OAuth2UserService;
+import nextstep.auth.oauth2.dto.OAuth2User;
+import nextstep.auth.oauth2.dto.OAuth2UserRequest;
+import nextstep.auth.oauth2.service.OAuth2UserService;
 import nextstep.member.entity.CustomOAuth2User;
 import nextstep.member.entity.Member;
 import nextstep.member.repository.MemberRepository;
